refactor: Move component interfaces to root types file
