+++ test/mocks/components/index.ts
@@ -4,3 +4,5 @@ export * from './pubsub'
 export * from './pg'
 export * from './config'
 export * from './uws'
+export * from './redis'
+export * from './archipelago-stats'
