+++ test/mocks/components/logs.ts
@@ -1,6 +1,4 @@
-import { IMetricsComponent } from '@well-known-components/interfaces'
 import { ILoggerComponent } from '@well-known-components/interfaces/dist/components/logger'
-import { createLogComponent } from '@well-known-components/logger'
 
 export const mockLogs: jest.Mocked<ILoggerComponent> = {
   getLogger: jest.fn().mockReturnValue({
